+++ packages/design-system/src/components/ButtonIcon/index.ts
@@ -1,4 +1,4 @@
-import ButtonIcon from './variations/ButtonIcon';
+import { ButtonIcon } from './variations/ButtonIcon';
 import ButtonIconToggle from './variations/ButtonIconToggle';
 import ButtonIconFloating from './variations/ButtonIconFloating';