+++ packages/design-system/src/components/InlineMessage/Primitive/InlineMessagePrimitive.tsx
@@ -64,4 +64,6 @@ const InlineMessagePrimitive = forwardRef(
 	},
 );
 
+InlineMessagePrimitive.displayName = 'InlineMessagePrimitive';
+
 export default InlineMessagePrimitive;