+++ packages/design-system/src/components/Loading/index.ts
@@ -1,3 +1,1 @@
-import Loading from './Loading';
-
-export default Loading;
+export * from './Loading';