+++ packages/design-system/src/stories/messaging/Popover.stories.tsx
@@ -1,9 +1,15 @@
 import { forwardRef, Ref } from 'react';
 
 import { action } from '@storybook/addon-actions';
-import { PopoverStateReturn } from 'reakit/ts';
-import { Popover, ButtonPrimary, ButtonIcon, StackVertical, Form } from '../../';
-import { PopoverDisclosureHTMLProps } from 'reakit';
+import {
+	Popover,
+	PopoverProps,
+	PopoverStateReturn,
+	ButtonPrimary,
+	ButtonIcon,
+	StackVertical,
+	Form,
+} from '../../';
 
 export default {
 	component: Popover,
@@ -13,7 +19,7 @@ export default {
 const EasyPopover = () => <StackVertical gap="S">Hello hello</StackVertical>;
 
 /* eslint-disable-next-line react/display-name */
-const OpenPopover = forwardRef((props: PopoverDisclosureHTMLProps, ref: Ref<HTMLButtonElement>) => {
+const OpenPopover = forwardRef((props: PopoverProps, ref: Ref<HTMLButtonElement>) => {
 	return (
 		<ButtonPrimary onClick={action('Clicked disclosure')} {...props} ref={ref}>
 			Open popover