+++ packages/faceted-search/src/components/FacetedToolbar/FacetedToolbar.component.js
@@ -1,8 +1,10 @@
 import PropTypes from 'prop-types';
+
 import { Form } from '@talend/design-system';
 import { getTheme } from '@talend/react-components/lib/theme';
 
 import { FACETED_MODE, USAGE_TRACKING_TAGS } from '../../constants';
+
 import cssModule from './FacetedToolbar.module.scss';
 
 const theme = getTheme(cssModule);
@@ -11,6 +13,7 @@ const SwitchFacetedMode = ({ facetedMode, onChange, t }) => (
 	<Form className={theme('tc-faceted-switch-mode')}>
 		<Form.ToggleSwitch
 			label={t('FACETED_SEARCH_QUERY', 'Query')}
+			checked={facetedMode === FACETED_MODE.ADVANCED}
 			onChange={() =>
 				onChange(facetedMode === FACETED_MODE.BASIC ? FACETED_MODE.ADVANCED : FACETED_MODE.BASIC)
 			}