+++ packages/dataviz/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 4.0.0
+
+### Major Changes
+
+- 96d688489: React: Upgrade to react 18 and @types/react 18
+- 4044f6198: ARCH-662 - Bump i18next from v20 to v23 and react-i18next from v11 to v13
+
+### Patch Changes
+
+- Updated dependencies [96d688489]
+- Updated dependencies [9a0732ec5]
+- Updated dependencies [da4cf2971]
+- Updated dependencies [4044f6198]
+  - @talend/react-components@12.0.0
+
 ## 3.0.1
 
 ### Patch Changes